+++ registry/components/ui/menubar/index.ts
@@ -1,0 +1,8 @@
+export { default as Menubar } from "./menubar.svelte";
+export { default as MenubarMenu } from "./menubar-menu.svelte";
+export { default as MenubarTrigger } from "./menubar-trigger.svelte";
+export { default as MenubarContent } from "./menubar-content.svelte";
+export { default as MenubarItem } from "./menubar-item.svelte";
+export { default as MenubarSeparator } from "./menubar-separator.svelte";
+export { default as MenubarLabel } from "./menubar-label.svelte";
+export { default as MenubarShortcut } from "./menubar-shortcut.svelte";